requirements.txt
dj-database-url==0.5.0
certifi==2019.11.28
chardet==3.0.4
Django==2.2.7
django-cors-headers==3.2.0
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
idna==2.8
Pillow==6.2.1
psycopg2==2.8.4
python-dotenv==0.10.3
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
waitress==1.3.1
whitenoise==4.1.4
urllib3==1.25.7
